+++ src/hr/fer/zemris/structures/dot/Functions.java
@@ -7,10 +7,10 @@ import java.util.Map;
 
 public class Functions 
 {
-	public static List<Integer> intersection(List<Integer> list1, List<Integer> list2)
+	public static List<Long> intersection(List<Long> list1, List<Long> list2)
 	{
-		List<Integer> resultList = new ArrayList<>();
-		Map<Integer, Boolean> map = new HashMap<Integer,Boolean>();
+		List<Long> resultList = new ArrayList<>();
+		Map<Long, Boolean> map = new HashMap<Long,Boolean>();
 		for(int i=0, len=list1.size();i<len;++i)
 			map.put(list1.get(i), true);
